+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Descriptive changelog
 (All dates are DD-MM-YYYY)
+##### Adding modular integer support, changing api
+- Added modular integers
+- Renamed invPrecise() bigInteger method to not()
+- renamed numberOfDigits() to numberOfDecimalDigits()
+- introduced BigNumber and BitwiseOperations interfaces 
+
+
 ##### Minor update - 0.0.8 - 02.04.2019
 - Released ios (for X64 and arm) and macos X64 artifacts
 - No functional changes